Remove serde dependency (#796)
* Remove dependency on serde and rely on mls_codec for serialization internally

Full removal of serde

Fixup

Iterator fixups

more changes

make iter public

* Reduce size of data collection serialization

* Fix formatting

* CI Fixes

* Fixup CI

* Fix after merge

* fixes after merge

* no_std build fixes

* remove empty file

* Update aws-mls/src/group/message_verifier.rs


* Remove empty file

---------

+++ aws-mls/benches/decap.rs
@@ -34,7 +34,7 @@ fn decap_setup(c: &mut Criterion) {
 fn bench_decap(
     bench_group: &mut BenchmarkGroup<WallTime>,
     cipher_suite: CipherSuite,
-    map: HashMap<usize, TestCase>,
+    map: HashMap<u32, TestCase>,
     added_leaves: &[LeafIndex],
     capabilities: Option<Capabilities>,
     extensions: Option<ExtensionList>,
